Added code to prevent crash when no items plottable
When there are no items plottable (may be the result of no datasets
provided or no x and y data provided), xDiff or yDiff may be infinity and
the code would crash when they're divided by xInterval or yInterval.
Checks have been added to only undergo these processing when the values
are finite.